[code-infra] Convert @mui/styled-engine-sc to TypeScript
Same setup as @mui/styled-engine (#48544) and @mui/private-theming
(#48565): hand-written `.js` + `.d.ts` source replaced with a single
`.ts`/`.tsx`, declarations now emitted by `tsc` instead of copied,
`--skipTsc` dropped from the build script, `tsconfig.build.json` added.

This package's hand-written `.d.ts` over-claimed the type surface by
re-exporting the entire `styled-components` named-export set via
`export * from 'styled-components'`, while the runtime `.js` only
re-exported `ThemeContext`, `keyframes`, `css`. To keep the exported
type surface 100% identical (proven via a bidirectional
type-equivalence probe over root + every subpath entry), the new
`src/index.ts` keeps the `export * from 'styled-components'` — so the
runtime JS now also re-exports the broader surface, fixing the prior
type/runtime divergence. The local default stays the dev-mode wrapper,
cast to `typeof scStyled` so its public type matches the hand `.d.ts`'s
`export { default } from 'styled-components'`.

Other JS deltas are limited to the standard true-TS conversion
artifacts: propTypes assignments now wrap in the
`process.env.NODE_ENV !== 'production' ? … : void 0` guard via the
`/* remove-proptypes */` marker, and `index`'s default export form
becomes `function styled; export default styled;` (same runtime).
`.d.ts` differences are limited to type-preserving forms (`import type`,
`export type *`, `declare`, expando-free default emit).

There are no internal MUI workspace consumers of `@mui/styled-engine-sc`,
so no downstream `tsconfig.build.json` `references` updates were needed.

## Published artifact diff

Diff of the published package before/after this PR: [`@mui/styled-engine-sc`].
