--- packages/mui-styled-engine-sc/src/GlobalStyles/GlobalStyles.d.ts
@@ -1,11 +1,0 @@
-import * as React from 'react';
-import { CSSObject, StyleFunction } from 'styled-components';
-
-export interface GlobalStylesProps<Theme extends object = {}> {
-  defaultTheme?: object | undefined;
-  styles: string | CSSObject | StyleFunction<Theme>;
-}
-
-export default function Global<Theme extends object = {}>(
-  props: GlobalStylesProps<Theme>,
-): React.JSX.Element;
