--- packages/mui-styled-engine-sc/src/GlobalStyles/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from './GlobalStyles';
-export * from './GlobalStyles';
