+++ packages/mui-styled-engine-sc/src/GlobalStyles/index.ts
@@ -1,1 +1,2 @@
 export { default } from './GlobalStyles';
+export type * from './GlobalStyles';
