--- packages/mui-styled-engine-sc/src/StyledEngineProvider/StyledEngineProvider.d.ts
@@ -1,8 +1,0 @@
-import * as React from 'react';
-
-export interface StyledEngineProviderProps {
-  children?: React.ReactNode;
-  injectFirst?: boolean | undefined;
-}
-
-export default function StyledEngineProvider(props: StyledEngineProviderProps): React.JSX.Element;
