--- packages/mui-styled-engine-sc/src/StyledEngineProvider/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from './StyledEngineProvider';
-export * from './StyledEngineProvider';
