+++ packages/mui-styled-engine-sc/src/StyledEngineProvider/index.ts
@@ -1,1 +1,2 @@
 export { default } from './StyledEngineProvider';
+export type * from './StyledEngineProvider';
