+++ packages/mui-styled-engine-sc/tsconfig.json
@@ -1,7 +1,9 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "skipLibCheck": true
+    "allowJs": false,
+    "skipLibCheck": true,
+    "types": ["react", "vitest/globals", "@mui/internal-code-infra/build-env"]
   },
   "include": ["src/**/*"]
 }
